+++ scripts/writeEnv.cjs
@@ -1,7 +1,7 @@
 // @ts-check
 /* eslint-disable @typescript-eslint/no-require-imports */
 
-const { version } = require("../../package.json");
+const { version } = require("../package.json");
 
 const { writeFileSync } = require("fs");
 const { resolve } = require("path");
@@ -10,7 +10,7 @@ const ENV = {
   PUBLIC_VERSION: version,
 };
 
-const ENV_PATH = resolve(__dirname, "../../.env");
+const ENV_PATH = resolve(__dirname, "../.env");
 
 const envEntries = Object.entries(ENV);
 let envString = "";